Merged Trond's commit from earlier today with this one. His commit message was:
Starting work on the smn keyboard: New desktop keyboard + revised mobile keyboard (added hyphen on top level, thanks to ML for feedback).

The actual changes in this commit compared to the previous revision are:
* changed internalName
* added desktop keyboard layout (based on old sme version, must be replaced)
* added hyphen & underscore to mobile keyboard
* added a couple of more displayNames

One revert compared to Trond's mobile changes: that one had capital ENG instead of Â as the upper right letter. I reverted that - it looks like a typo (we lost Â, and ENG is already found elsewhere).

Except for this little change (ENG back to Â) the keyboard layout file is now back to Trond's version, but now with the whole svn history attached to it.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/keyboards/smn@153503 c7155fb1-f0a7-4240-a2fc-2600b6f42f90